requirements.txt
aiogram~=2.20
alembic~=1.8.0
environs~=9.5.0
SQLAlchemy~=1.4.37
pycld3~=0.22.0